# backend/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Consultation API",
      "version": "1.0.0",
      "description": "API for managing medical consultations"
    },
    "servers": [
      {
        "url": "http://localhost:5000",
        "description": "Local development server"
      }
    ],
    "paths": {
...
      "/api/billing": {
      "post": {
        "summary": "Create a new billing record",
        "description": "Creates a billing record for a provider and sends email notification",
        "tags": ["Billing"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["provider_id", "consultation_id", "duration_minutes", "billing_amount", "date_range"],
                "properties": {
                  "provider_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "consultation_id": {
                    "type": "integer",
                    "example": 101
                  },
                  "duration_minutes": {
                    "type": "integer",
                    "example": 30
                  },
                  "billing_amount": {
                    "type": "number",
                    "format": "float",
                    "example": 150.50
                  },
                  "date_range": {
                    "type": "string",
                    "example": "June 2024"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Billing record created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Billing record created, notification sent via email."
                    },
                    "billing": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "provider_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "consultation_id": {
                          "type": "integer",
                          "example": 101
                        },
                        "duration_minutes": {
                          "type": "integer",
                          "example": 30
                        },
                        "billing_amount": {
                          "type": "number",
                          "format": "float",
                          "example": 150.50
                        },
                        "date_range": {
                          "type": "string",
                          "example": "June 2024"
                        }
                      }
                    },
                    "notification": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "provider_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "type": {
                          "type": "string",
                          "example": "billing"
                        },
                        "message": {
                          "type": "string",
                          "example": "Billing summary for June 2024 has been generated. Total: $150.50."
                        },
                        "status": {
                          "type": "string",
                          "example": "unread"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Provider not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Provider not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "summary": "Get all billing records",
        "description": "Retrieves a list of all billing records",
        "tags": ["Billing"],
        "responses": {
          "200": {
            "description": "List of billing records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "provider_id": {
                        "type": "integer",
                        "example": 1
                      },
                      "consultation_id": {
                        "type": "integer",
                        "example": 101
                      },
                      "duration_minutes": {
                        "type": "integer",
                        "example": 30
                      },
                      "billing_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 150.50
                      },
                      "date_range": {
                        "type": "string",
                        "example": "June 2024"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/billing/{provider_id}": {
      "get": {
        "summary": "Get billing records by provider ID",
        "description": "Retrieves all billing records for a specific provider",
        "tags": ["Billing"],
        "parameters": [
          {
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the provider"
          }
        ],
        "responses": {
          "200": {
            "description": "List of billing records for the provider",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "provider_id": {
                        "type": "integer",
                        "example": 1
                      },
                      "consultation_id": {
                        "type": "integer",
                        "example": 101
                      },
                      "duration_minutes": {
                        "type": "integer",
                        "example": 30
                      },
                      "billing_amount": {
                        "type": "number",
                        "format": "float",
                        "example": 150.50
                      },
                      "date_range": {
                        "type": "string",
                        "example": "June 2024"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/billing/{id}": {
      "put": {
        "summary": "Update a billing record",
        "description": "Updates a billing record and notifies the provider",
        "tags": ["Billing"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the billing record"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "duration_minutes": {
                    "type": "integer",
                    "example": 45
                  },
                  "billing_amount": {
                    "type": "number",
                    "format": "float",
                    "example": 200.00
                  },
                  "date_range": {
                    "type": "string",
                    "example": "June 2024"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Billing record updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Billing record updated and notification sent."
                    },
                    "billing": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "provider_id": {
                          "type": "integer",
                          "example": 1
                        },
                        "consultation_id": {
                          "type": "integer",
                          "example": 101
                        },
                        "duration_minutes": {
                          "type": "integer",
                          "example": 45
                        },
                        "billing_amount": {
                          "type": "number",
                          "format": "float",
                          "example": 200.00
                        },
                        "date_range": {
                          "type": "string",
                          "example": "June 2024"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Billing record or provider not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Delete a billing record",
        "description": "Deletes a billing record and notifies the provider",
        "tags": ["Billing"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the billing record to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Billing record deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Billing record deleted and email notification sent."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Billing record or provider not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    },
    "components": {
...
      "securitySchemes": {}
    }
  }